ml_services/interview-llm-service/requirements.txt
# FastAPI Server
fastapi
uvicorn[standard]
python-dotenv

# Core ML/LLM Libraries
torch
transformers
accelerate
bitsandbytes
pydantic
scikit-learn

# Hugging Face Hub for model downloading
huggingface-hub>=0.23.0